Move socket/plug stuff into sockets.ts
